messaging_app/messaging_app/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.4; python_version >= '3.10'
django-filter==24.3; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-nested-routers==0.94.1; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
mysqlclient==2.2.1 